separation des modeles en fichiers

+++ modeles/fenetre.py
@@ -8,13 +8,16 @@
 """
 __author__ = 'Daniel-Junior Dubé & Sarah Laflamme'
 
-from modeles.theme import *
-from modeles.image import *
 from sqlalchemy import *
+from modeles.base import Base
+from modeles.image import Image
+from modeles.theme import Theme
+from modeles.zone import Zone
+from modeles.zone_base import ZoneBase
+from modeles.zone_image import ZoneImage
+from modeles.zone_video import ZoneVideo
+from modeles.zone_table import ZoneTable
 from sqlalchemy.orm import relationship, backref, sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
-
-Base = declarative_base()
 
 class Fenetre(Base):
     """
@@ -44,19 +47,19 @@ class Fenetre(Base):
     id_image_fond = Column(
         Integer, 
         ForeignKey('Images.id', onupdate="cascade", ondelete="set default"), 
-        default=1
-        )    # DEFAULT VALIDE??
+        nullable=True
+        ) 
     id_theme = Column(
         Integer, 
         ForeignKey('Themes.id', onupdate="cascade", ondelete="set default"), 
         default=1
-        )  # DEFAULT VALIDE??
+        ) 
     image_fond = relationship(
-        Image, 
+        "Image", 
         foreign_keys=[id_image_fond]
         )
     theme = relationship(
-        Theme, 
+        "Theme", 
         backref=backref(
             'fenetres', 
             uselist=True),
@@ -72,10 +75,13 @@ class Fenetre(Base):
         data = dict(
             id = self.id,
             nom = self.nom,
-            couleur_fond = self.couleur_fond,
-            image_fond = self.image_fond.serialiser_en_json(),
-            theme = self.theme.serialiser_en_json()
+            theme = self.theme.serialiser_en_json(),
+            couleur_fond = self.couleur_fond
             )
+        if self.image_fond != None:
+            data['image_fond'] = self.image_fond.serialiser_en_json()
+        else:
+            data['image_fond'] = "undefined"
         for zone in self.zones:
             zones_data.append(zone.serialiser_en_json())
         data['zones'] = zones_data
@@ -92,34 +98,36 @@ class Fenetre(Base):
                 data (Dict) : Dictionnaire qui contient les valeurs à assigner.
         """
         if data.get('nom') != None : self.nom = data['nom']
+        if data.get('image_fond') == None or data.get('image_fond') == "undefined" : self.id_image_fond = "undefined"
+        else: self.image_fond = session.query(Image).filter(Image.id == data['image_fond']['id']).one()
         if data.get('couleur_fond') != None : self.couleur_fond = data['couleur_fond']
-        if (self.id_theme != data['theme']['id']):
-            self.theme = session.query(Theme).filter(Theme.id == data['theme']['id']).one()
-        if (self.id_image_fond != data['image_fond']['id']):
-            self.image_fond = session.query(Image).filter(Image.id == data['image_fond']['id']).one()
-        for zone in data['zones']:
-            if zone['id'] == 0:
-                if zone['type'] == 'ZoneBase':
-                    nouvelle_zone = ZoneBase(fenetre=self)
-                    nouvelle_zone.deserialiser_de_json(session, zone)
-                    session.add(nouvelle_zone)
-                elif zone['type'] == 'ZoneTable':
-                    nouvelle_zone = ZoneTable(fenetre=self)
-                    nouvelle_zone.deserialiser_de_json(session, zone)
-                    session.add(nouvelle_zone)
-                elif zone['type'] == 'ZoneImage':
-                    nouvelle_zone = ZoneImage(fenetre=self)
-                    nouvelle_zone.deserialiser_de_json(session, zone)
-                    session.add(nouvelle_zone)
-                elif zone['type'] == 'ZoneVideo':
-                    nouvelle_zone = ZoneVideo(fenetre=self)
-                    nouvelle_zone.deserialiser_de_json(session, zone)
-                    session.add(nouvelle_zone)
+        if data.get('theme') != None :
+            if (self.id_theme != data['theme']['id']):
+                self.theme = session.query(Theme).filter(Theme.id == data['theme']['id']).one()
+        if data.get('zones') != None :
+            for zone in data['zones']:
+                if zone['id'] == 0:
+                    if zone['type'] == 'ZoneBase':
+                        nouvelle_zone = ZoneBase(fenetre=self)
+                        nouvelle_zone.deserialiser_de_json(session, zone)
+                        session.add(nouvelle_zone)
+                    elif zone['type'] == 'ZoneTable':
+                        nouvelle_zone = ZoneTable(fenetre=self)
+                        nouvelle_zone.deserialiser_de_json(session, zone)
+                        session.add(nouvelle_zone)
+                    elif zone['type'] == 'ZoneImage':
+                        nouvelle_zone = ZoneImage(fenetre=self)
+                        nouvelle_zone.deserialiser_de_json(session, zone)
+                        session.add(nouvelle_zone)
+                    elif zone['type'] == 'ZoneVideo':
+                        nouvelle_zone = ZoneVideo(fenetre=self)
+                        nouvelle_zone.deserialiser_de_json(session, zone)
+                        session.add(nouvelle_zone)
+                    else:
+                        print("Type de zone inexistante!") # IMPLÉMENTER UNE ERREUR CORRECTE
+                elif zone['id'] > 0:
+                    session.query(Zone).filter(Zone.id == zone['id']).one().deserialiser_de_json(session, zone)
+                elif zone['id'] < 0:
+                    session.delete(session.query(Zone).filter(Zone.id == -zone['id']).one())
                 else:
-                    print("Type de zone inexistante!") # IMPLÉMENTER UNE ERREUR CORRECTE
-            elif zone['id'] > 0:
-                session.query(Zone).filter(Zone.id == zone['id']).one().deserialiser_de_json(session, zone)
-            elif zone['id'] < 0:
-                session.delete(session.query(Zone).filter(Zone.id == -zone['id']).one())
-            else:
-                print('Impossible de déserialiser la zone')
+                    print('Impossible de déserialiser la zone')
